Serialization improvements and bug fix

+++ HouseManagement/src/Aufgabe1PLC21WS/EigentumsWohnung.java
@@ -1,10 +1,10 @@
 package Aufgabe1PLC21WS;
 
 public class EigentumsWohnung extends Wohnung {
-    private Integer betriebskosten;
-    private Integer reparaturruecklage;
+    private Double betriebskosten;
+    private Double reparaturruecklage;
 
-    public EigentumsWohnung(Integer id, Double flaeche, Integer zimmeranzahl, Integer stockwerk, Integer baujahr, Integer PLZ, String strasse, Integer hausnummer, Integer top, Integer betriebskosten, Integer reparaturruecklage) {
+    public EigentumsWohnung(Integer id, Double flaeche, Integer zimmeranzahl, Integer stockwerk, Integer baujahr, Integer PLZ, String strasse, Integer hausnummer, Integer top, Double betriebskosten, Double reparaturruecklage) {
         super(id, flaeche, zimmeranzahl, stockwerk, baujahr, PLZ, strasse, hausnummer, top);
         this.betriebskosten = betriebskosten;
         this.reparaturruecklage = reparaturruecklage;
